hboxcontainer moment
just added an hboxcontainer in a scroll container so that you can easily scroll through the participants, as rn more than like 5 dont properly fit on the screen unless you resize. wrote a bit of code to compensate for the problem of the characters now become children  of panel nodes.