More cast maker changes, started map maker
Map maker implementation is still extremely early and does not work as intended.

Also, the relationship editor in the cast maker is still a bit wonky. Aside from that though, I added the options to clear the current character list and to bulk add characters. Also added a temporary logo for shits and giggles

I added music for the cast maker but it's set to not autoplay and loop is turned off, so make sure you enable that if you want to hear it.