Created item editor for character
Created a new scene which houses the item option button and a button to delete itself.

Also, added a button on FighterCreator that acts similarly to the relationship editor, but instead it pulls up a window editing the characters inventory. In here, pressing the "add item" button just instantiates a new instance of the ItemEditor scene.

Added a section in character_info to popup the item window and set the title of the window to have the character's name. Similar to the relationship editor yet again.