Started implementing logic for linking areas
Doesnt fully work yet but it's recognized as which button is the initiator and adds itself to a group called "CurrentlyLinking", and upon selecting an area it adds them to the same group too. The idea is to get both of these nodes, check them both for self, and then append whats left to that area's links.